requirements.txt
certifi==2017.4.17
chardet==3.0.4
future==0.16.0
gspread==0.6.2
httplib2==0.10.3
idna==2.5
oauth2client==4.1.2
oauthlib==2.0.2
pkg-resources==0.0.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
python-twitter==3.3
requests==2.18.1
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
